Split test_attack_telem_classes.py and test_technique_telems.py into separate test files
