+++ monkey/infection_monkey/telemetry/tests/attack/test_t1129_telem.py
@@ -1,0 +1,22 @@
+import pytest
+
+from common.utils.attack_utils import ScanStatus, UsageEnum
+from infection_monkey.telemetry.attack.t1129_telem import T1129Telem
+
+
+STATUS = ScanStatus.USED
+USAGE = UsageEnum.SMB
+
+
+@pytest.fixture
+def T1129_telem_test_instance():
+    return T1129Telem(STATUS, USAGE)
+
+
+def test_T1129_send(T1129_telem_test_instance, spy_send_telemetry):
+    T1129_telem_test_instance.send()
+    expected_data = {'status': STATUS.value,
+                     'technique': 'T1129',
+                     'usage': USAGE.name}
+    assert spy_send_telemetry.data == expected_data
+    assert spy_send_telemetry.telem_category == 'attack'
